refactor(pickers): extract shared picker support (#527)
Problem:  still owned the shared picker plumbing for
placeholder rows, scoped cache keys, pagination, cache patching, and
common action helpers. That made later picker surface splits depend on a
single giant module.

Solution: move the shared picker support layer into
 and keep  focused on the concrete
picker surfaces. Update the picker specs to clear the new shared module
between runs.
